First pass on tamago/noos
Disable usage of the "testing" package in actual runtime code.
