+++ dialog/file_tamago.go
@@ -1,0 +1,9 @@
+//go:build tamago || noos
+
+package dialog
+
+import "fyne.io/fyne/v2"
+
+func getFavoriteLocations() (map[string]fyne.ListableURI, error) {
+	return map[string]fyne.ListableURI{}, nil
+}
