+++ test/app.go
@@ -4,7 +4,6 @@ package test // import "fyne.io/fyne/v2/test"
 import (
 	"net/url"
 	"sync"
-	"testing"
 
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/internal"
@@ -150,15 +149,6 @@ func (a *app) transitionCloud(p fyne.CloudProvider) {
 
 	// after transition ensure settings listener is fired
 	a.settings.apply()
-}
-
-// NewTempApp returns a new dummy app and tears it down at the end of the test.
-//
-// Since: 2.5
-func NewTempApp(t testing.TB) fyne.App {
-	app := NewApp()
-	t.Cleanup(func() { NewApp() })
-	return app
 }
 
 // NewApp returns a new dummy app used for testing.
