+++ test/app_test.go
@@ -3,8 +3,9 @@ package test
 import (
 	"testing"
 
-	"fyne.io/fyne/v2"
 	"github.com/stretchr/testify/assert"
+
+	"fyne.io/fyne/v2"
 )
 
 func TestTestApp_CloudProvider(t *testing.T) {
