+++ test/notification_test.go
@@ -1,10 +1,9 @@
-package test_test
+package test
 
 import (
 	"testing"
 
 	"fyne.io/fyne/v2"
-	"fyne.io/fyne/v2/test"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -13,14 +12,14 @@ func TestAssertNotificationSent(t *testing.T) {
 	n := fyne.NewNotification("Test Title", "Some content")
 	myApp := fyne.CurrentApp()
 
-	test.AssertNotificationSent(t, n, func() {
+	AssertNotificationSent(t, n, func() {
 		fyne.CurrentApp().SendNotification(n)
 	})
 	assert.Equal(t, myApp, fyne.CurrentApp())
 }
 
 func TestAssertNotificationSent_Nil(t *testing.T) {
-	test.AssertNotificationSent(t, nil, func() {
+	AssertNotificationSent(t, nil, func() {
 		// don't send anything
 	})
 }
@@ -28,7 +27,7 @@ func TestAssertNotificationSent_Nil(t *testing.T) {
 func TestAssertNotificationSent_NotSent(t *testing.T) {
 	tt := &testing.T{}
 
-	test.AssertNotificationSent(tt, &fyne.Notification{}, func() {
+	AssertNotificationSent(tt, &fyne.Notification{}, func() {
 		// don't send anything
 	})
 	assert.True(t, tt.Failed(), "notification assert should fail if no notification was sent")
